% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.survivalReport.R
\name{rs.survivalReport}
\alias{rs.survivalReport}
\title{Returns a survival report}
\usage{
rs.survivalReport()
}
\value{
a report dataframe
}
\description{
this function is similar survival report
}
\details{
this function is similar survival report
}
\author{
Wheater, S.
}
